requirements.txt
aiogram
selenium
beautifulsoup4
webdriver-manager
pandas
fake-useragent
requests
python-dotenv 